OK-674: Jaettu henkilön valintalaskentaan liittyviä komponentteja eri tiedostoihin
