+++ src/app/haku/[oid]/henkilo/[hakemusOid]/components/henkilon-valintalaskenta.tsx
@@ -1,0 +1,200 @@
+'use client';
+
+import { useTranslations } from '@/app/hooks/useTranslations';
+import { Divider, Stack, styled, Typography } from '@mui/material';
+import { OphButton, OphTypography } from '@opetushallitus/oph-design-system';
+import { withDefaultProps } from '@/app/lib/mui-utils';
+import {
+  LaskentaActorRef,
+  LaskentaEvent,
+  LaskentaEventType,
+  LaskentaMachineSnapshot,
+  LaskentaState,
+  useLaskentaError,
+  useLaskentaState,
+} from '@/app/lib/state/laskenta-state';
+import { HenkilonHakukohdeTuloksilla } from '../lib/henkilo-page-types';
+import useToaster from '@/app/hooks/useToaster';
+import { HaunAsetukset } from '@/app/lib/types/haun-asetukset';
+import { Haku } from '@/app/lib/types/kouta-types';
+import { ErrorAlert } from '@/app/components/error-alert';
+import { useSelector } from '@xstate/react';
+import { SeurantaTiedot } from '@/app/lib/types/laskenta-types';
+import { TFunction } from 'i18next';
+import { ProgressBar } from '@/app/components/progress-bar';
+import { SuorittamattomatHakukohteet } from './suorittamattomat-hakukohteet';
+import { ConfirmationModalDialog } from './confirmation-modal-dialog';
+
+const LaskentaButton = withDefaultProps(
+  styled(OphButton)({
+    alignSelf: 'flex-start',
+  }),
+  {
+    variant: 'contained',
+  },
+);
+
+const LaskentaStateButton = ({
+  state,
+  send,
+}: {
+  state: LaskentaMachineSnapshot;
+  send: (event: LaskentaEvent) => void;
+}) => {
+  const { t } = useTranslations();
+
+  switch (true) {
+    case state.hasTag('stopped') && !state.hasTag('completed'):
+      return (
+        <LaskentaButton
+          key="suorita"
+          onClick={() => {
+            send({ type: LaskentaEventType.START });
+          }}
+        >
+          {t('henkilo.suorita-valintalaskenta')}
+        </LaskentaButton>
+      );
+    case state.hasTag('started'):
+      return (
+        <LaskentaButton
+          key="keskeyta"
+          variant="outlined"
+          disabled={state.hasTag('canceling')}
+          onClick={() => {
+            send({ type: LaskentaEventType.CANCEL });
+          }}
+        >
+          {t('henkilo.keskeyta-valintalaskenta')}
+        </LaskentaButton>
+      );
+    case state.hasTag('completed'):
+      return (
+        <LaskentaButton
+          key="sulje"
+          variant="outlined"
+          onClick={() => {
+            send({ type: LaskentaEventType.RESET_RESULTS });
+          }}
+        >
+          {t('henkilo.sulje-laskennan-tiedot')}
+        </LaskentaButton>
+      );
+    default:
+      return null;
+  }
+};
+
+const getLaskentaStatusText = (
+  state: LaskentaMachineSnapshot,
+  seurantaTiedot: SeurantaTiedot | null,
+  t: TFunction,
+) => {
+  switch (true) {
+    case state.hasTag('canceling') ||
+      (state.matches(LaskentaState.FETCHING_SUMMARY) &&
+        state.context.seurantaTiedot?.tila === 'PERUUTETTU'):
+      return `${t('henkilo.keskeytetaan-laskentaa')} `;
+    case state.matches(LaskentaState.STARTING) ||
+      (state.hasTag('started') && seurantaTiedot == null):
+      return `${t('henkilo.kaynnistetaan-laskentaa')} `;
+    case state.hasTag('started'):
+      return seurantaTiedot?.jonosija
+        ? `${'henkilo.tehtava-on-laskennassa-jonosijalla'} ${seurantaTiedot?.jonosija}. `
+        : `${t('henkilo.tehtava-on-laskennassa-parhaillaan')}. `;
+    case state.hasTag('completed'):
+      return `${t('henkilo.laskenta-on-paattynyt')}. `;
+    default:
+      return '';
+  }
+};
+
+const LaskentaResult = ({ actorRef }: { actorRef: LaskentaActorRef }) => {
+  const { t } = useTranslations();
+
+  const laskentaError = useLaskentaError(actorRef);
+
+  const state = useSelector(actorRef, (s) => s);
+  const seurantaTiedot = useSelector(actorRef, (s) => s.context.seurantaTiedot);
+
+  const laskentaPercent = seurantaTiedot
+    ? Math.round(
+        (100 *
+          (seurantaTiedot?.hakukohteitaValmiina +
+            seurantaTiedot?.hakukohteitaKeskeytetty)) /
+          seurantaTiedot?.hakukohteitaYhteensa,
+      )
+    : 0;
+
+  switch (true) {
+    case state.matches({ [LaskentaState.IDLE]: LaskentaState.ERROR }):
+      return (
+        <ErrorAlert
+          title={t('henkilo.valintalaskenta-epaonnistui')}
+          message={laskentaError}
+        />
+      );
+    case state.hasTag('started') || state.hasTag('completed'):
+      return (
+        <>
+          <OphTypography variant="h4">
+            {t('henkilo.valintalaskenta')}
+          </OphTypography>
+          <ProgressBar value={laskentaPercent} />
+          <Typography>
+            {getLaskentaStatusText(state, seurantaTiedot, t)}
+            {seurantaTiedot &&
+              `${t('henkilo.hakukohteita-valmiina')} ${seurantaTiedot.hakukohteitaValmiina}/${seurantaTiedot.hakukohteitaYhteensa}. ` +
+                `${t('henkilo.suorittamattomia-hakukohteita')} ${seurantaTiedot?.hakukohteitaKeskeytetty ?? 0}.`}
+          </Typography>
+        </>
+      );
+    default:
+      return null;
+  }
+};
+
+export const HenkilonValintalaskenta = ({
+  haku,
+  haunAsetukset,
+  hakukohteet,
+}: {
+  haku: Haku;
+  haunAsetukset: HaunAsetukset;
+  hakukohteet: Array<HenkilonHakukohdeTuloksilla>;
+}) => {
+  const { addToast } = useToaster();
+
+  const [state, send, actorRef] = useLaskentaState({
+    haku,
+    haunAsetukset,
+    hakukohteet,
+    addToast,
+  });
+
+  return (
+    <Stack spacing={2}>
+      <ConfirmationModalDialog
+        open={state.matches(LaskentaState.WAITING_CONFIRMATION)}
+        onAnswer={(answer: boolean) => {
+          if (answer) {
+            send({ type: LaskentaEventType.CONFIRM });
+          } else {
+            send({ type: LaskentaEventType.CANCEL });
+          }
+        }}
+      />
+      <LaskentaResult actorRef={actorRef} />
+      <LaskentaStateButton state={state} send={send} />
+      {state.hasTag('completed') && (
+        <SuorittamattomatHakukohteet
+          actorRef={actorRef}
+          hakukohteet={hakukohteet}
+        />
+      )}
+      {(state.hasTag('started') || state.hasTag('completed')) && (
+        <Divider sx={{ paddingTop: 1 }} />
+      )}
+    </Stack>
+  );
+};
